interventions page + post
Hi! So, I created a page for the RC Interventions which also schows on the tiles and navmenu, I think that is all fine and pretty.

I also created a sample post for the interventions, as I thouhgt we will add them like this (each intervention = one post) but it seems that posts only and automaticlally appear in the activity page... Hope @duncdrum has some neat solution to that ;)