upload intro translation and panels
For Niqiu and The Watch.
The word files have the introduction and then the translation (with Chinese original). For the upload:
one separate page for the introduction, with a link to the translation (and back); the Chinese characters can probably be deleted as they are also in the panels. And the acknowledgement note maybe not as a footnote but after the translation. But I don't know what looks best.